requirements.txt
Click==7.0
datefinder==0.7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonpickle==1.2
MarkupSafe==1.1.1
numpy==1.17.4
opencv-python==4.1.2.30
pandas==0.24.2
Pillow==6.2.1
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
six==1.13.0
Werkzeug==0.16.0
